A few test tidy ups, including working with modern git
Some notes added on how to get the tests to pass.

Changed some of the gomplate temporary paths to gofsimpl paths
instead.

git init with recent git creates a `main` branch not a `master`
branch. This makes the test fail. As a simple workaround I've made
sure we're on `master`.